+++ airflow/dags/utils/dag_data.py
@@ -1,15 +1,13 @@
 from typing import List
 
 import requests
-from pydantic import BaseModel, TypeAdapter
+from pydantic import TypeAdapter
+
+from dags.utils.schemas import DagMetaData
 
 
-class DagMetaData(BaseModel):
-    project: str
-    keywords: str
-    notify: str
-
-
-response = requests.get('http://dag-info:3000')
-ta = TypeAdapter(List[DagMetaData])
-dags_metadata = ta.validate_python(response.json())
+def get_dag_metadata():
+    response = requests.get('http://dag-info:3000')
+    ta = TypeAdapter(List[DagMetaData])
+    dags_metadata = ta.validate_python(response.json())
+    return dags_metadata
