+++ airflow/dags/utils/schemas.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class DagMetaData(BaseModel):
+    project: str
+    keywords: str
+    notify: str
